backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doogie-chat"
version = "1.0.0"
description = "A chat bot application with a Hybrid RAG (BM25 + FAISS) and GraphRAG system"
authors = [
    {name = "Doogie Chat Team", email = "admin@example.com"}
]
readme = "../README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Web Framework
    "fastapi>=0.108.0",
    "uvicorn[standard]", # Allow latest version
    "websockets<14.0", # Pin below version with deprecation warning via uvicorn
    "aiohttp>=3.9.3",  # Updated for Python 3.12 compatibility
    
    # Database
    "sqlalchemy>=2.0.28",
    "alembic>=1.15.1",
    
    # Authentication
    "python-jose>=3.3.0", # Removed [cryptography] extra
    "cryptography>=42.0", # Added cryptography directly
    "passlib[bcrypt]>=1.7.4",
    "bcrypt~=3.2.0",  # Pinned to 3.2.x for passlib 1.7.4 compatibility
    "python-multipart>=0.0.9",
    
    # Data Validation
    "pydantic>=2.10.6",
    "pydantic-settings>=2.2.1",
    "email-validator>=2.1.1",
    
    # RAG Components
    # Using alternative vector store as faiss-cpu is difficult to build from source for Python 3.12
    "annoy>=1.17.3",  # Alternative to faiss-cpu for vector search
    "rank-bm25>=0.2.2",
    "networkx>=3.2.1",
    "sentence-transformers>=3.4.1",
    "scikit-learn>=1.4.1",
    
    # Document Processing
    "pypdf>=4.0.0",  # Replaced deprecated PyPDF2
    "python-docx>=1.1.0",
    "markdown>=3.6",
    "python-frontmatter>=1.1.0",
    "PyYAML>=6.0.1",
    
    # Utilities
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
    "loguru>=0.7.2",
    "requests>=2.32.0",
    "GitPython>=3.1.43",
    
    # LLM Clients
    "anthropic>=0.21.3",
    "google-generativeai>=0.4.0",

    # Docker
    "docker>=7.1.0",
    "mcp>=1.6.0", # Added MCP SDK dependency
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0", # Added for test coverage
    "black>=24.2.0",
    "pylint>=3.1.0",
    "isort>=5.13.2",
    "mypy>=1.9.0",
    "bandit>=1.7.7",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools]
# Explicitly tell setuptools which package directory to include
packages = ["app"]

[tool.black]
line-length = 120
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 120

[tool.pylint]
max-line-length = 120
disable = ["C0111", "C0103", "C0303", "W1203"]
